input chips: adjusted layout
