--- webapp/components/form/InputChips/Chip/index.js
@@ -1,1 +1,0 @@
-export { default } from './Chip'
